+++ docs/src/components/examples/CompositionInputFieldValibot.vue
@@ -5,7 +5,7 @@
 
 <script setup>
 import { useField } from 'vee-validate';
-import { toTypedSchema } from '@vee-validate/valibot';
+import { toTypedSchema } from '@@zaalbarxx/vee-validate-valibot';
 import { string, email, minLength } from 'valibot';
 
 const { value, errorMessage } = useField('email', toTypedSchema(string([minLength(1, 'Required'), email()])));